Creating the script to run a regression on the size distribution. This data set keeps the 'year' tagged to the data so that we can see if there is a year effect.
Already modified the 'tidy' function. The parameter 'year' essentially asks if you want to group the resulting data by year. It defaults to FALSE, which leaves out the year. TRUE will group by year.

For PISCO, the code to generate bb5 was modified to aggregate by year. This keeps the 'year' tag in.